Fix merge conflicts with refactored team code base
