+++ src/main/java/seedu/address/logic/commands/view/ViewCommand.java
@@ -1,5 +1,7 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.view;
 
+import seedu.address.logic.commands.Command;
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
