+++ src/main/java/seedu/address/logic/commands/view/ViewTaskListCommand.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.view;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.logic.parser.CliSyntax.VIEW_DEADLINE_LIST;
@@ -7,6 +7,7 @@ import static seedu.address.logic.parser.CliSyntax.VIEW_TASK_LIST;
 import static seedu.address.logic.parser.CliSyntax.VIEW_TODO_LIST;
 
 import seedu.address.commons.core.Messages;
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
