+++ src/main/java/seedu/address/logic/parser/EditCommandParser.java
@@ -19,7 +19,7 @@ import seedu.address.logic.commands.EditLoginCommand;
 import seedu.address.logic.commands.EditLoginCommand.EditLoginDescriptor;
 import seedu.address.logic.commands.EditStudentCommand;
 import seedu.address.logic.commands.EditStudentCommand.EditPersonDescriptor;
-import seedu.address.logic.commands.ViewCommand;
+import seedu.address.logic.commands.view.ViewCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.flag.Flag;
