+++ src/main/java/seedu/address/logic/parser/ViewCommandParser.java
@@ -17,14 +17,14 @@ import java.util.Optional;
 import java.util.logging.Logger;
 
 import seedu.address.commons.core.LogsCenter;
-import seedu.address.logic.commands.ViewAllStudentsCommand;
-import seedu.address.logic.commands.ViewCommand;
-import seedu.address.logic.commands.ViewMissionDeadlineCommand;
-import seedu.address.logic.commands.ViewOneStudentCommand;
-import seedu.address.logic.commands.ViewQuestDeadlineCommand;
-import seedu.address.logic.commands.ViewTaskListCommand;
-import seedu.address.logic.commands.ViewUngradedMissionCommand;
-import seedu.address.logic.commands.ViewUngradedQuestCommand;
+import seedu.address.logic.commands.view.ViewAllStudentsCommand;
+import seedu.address.logic.commands.view.ViewCommand;
+import seedu.address.logic.commands.view.ViewMissionDeadlineCommand;
+import seedu.address.logic.commands.view.ViewOneStudentCommand;
+import seedu.address.logic.commands.view.ViewQuestDeadlineCommand;
+import seedu.address.logic.commands.view.ViewTaskListCommand;
+import seedu.address.logic.commands.view.ViewUngradedMissionCommand;
+import seedu.address.logic.commands.view.ViewUngradedQuestCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.flag.Flag;
 import seedu.address.model.student.Name;
