+++ src/test/java/seedu/address/logic/commands/ViewAllStudentsCommandTest.java
@@ -9,6 +9,7 @@ import static seedu.address.testutil.TypicalStudents.getTypicalAddressBook;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import seedu.address.logic.commands.view.ViewAllStudentsCommand;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.testutil.TypicalManagers;
