+++ src/test/java/seedu/address/logic/commands/ViewUpcomingConsultationsCommandTest.java
@@ -7,6 +7,7 @@ import static seedu.address.testutil.TypicalStudents.getTypicalAddressBook;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import seedu.address.logic.commands.view.ViewUpcomingConsultationsCommand;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.testutil.TypicalManagers;
